requirements.txt
annotated-types==0.5.0
anyio==3.7.1
black==23.7.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
dnspython==2.4.0
exceptiongroup==1.1.2
fastapi==0.78.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
iniconfig==2.0.0
mangum==0.17.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pika==1.3.2
platformdirs==3.9.1
pluggy==1.2.0
pydantic==1.10.12
pydantic_core==2.6.3
pyhere==1.0.0
pymongo==4.4.1
pytest==7.4.0
pytest-mock==3.11.1
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
sniffio==1.3.0
starlette==0.19.1
tomli==2.0.1
typeguard==4.1.5
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.23.0
